[DEMO-2521] Move footnote code into Summary
* Moved footnote validator into the Summary folder as FootnoteValidator.
* Moved code previously in Footnote::Orchestrator to just be part of
summary.
