+++ src/orchestrator.rb
@@ -3,7 +3,6 @@ require './src/provision/orchestrator'
 require './src/install/orchestrator'
 require './src/teardown/orchestrator'
 require './src/summary/orchestrator'
-require './src/footnote/orchestrator'
 
 class Orchestrator
 
@@ -18,7 +17,6 @@ class Orchestrator
     @install_orchestrator = install_orchestrator || Install::Orchestrator.new(context)
     @teardown_orchestrator = teardown_orchestrator || Teardown::Orchestrator.new(context)
     @summary_orchestrator = summary_orchestrator || Summary::Orchestrator.new(context)
-    @footnote_orchestrator = footnote_orchestrator || Footnote::Orchestrator.new(context)
   end
 
   def execute(arguments = ARGV)
@@ -28,8 +26,7 @@ class Orchestrator
     else
       @provision_orchestrator.execute()
       @install_orchestrator.execute()
-      @summary_orchestrator.execute()
-      return @footnote_orchestrator.execute()
+      return @summary_orchestrator.execute()
     end
   end
